Fix circular import with new modules
