+++ smartsim/entity/_mock.py
@@ -1,0 +1,58 @@
+# BSD 2-Clause License
+#
+# Copyright (c) 2021-2024, Hewlett Packard Enterprise
+# All rights reserved.
+#
+# Redistribution and use in source and binary forms, with or without
+# modification, are permitted provided that the following conditions are met:
+#
+# 1. Redistributions of source code must retain the above copyright notice, this
+#    list of conditions and the following disclaimer.
+#
+# 2. Redistributions in binary form must reproduce the above copyright notice,
+#    this list of conditions and the following disclaimer in the documentation
+#    and/or other materials provided with the distribution.
+#
+# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+# AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+# IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+# DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
+# FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+# DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+# SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+# CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
+# OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+# OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+
+"""This module contains stubs of functionality that is not currently
+implemented. 
+
+THIS WHOLE MODULE SHOULD BE REMOVED IN FUTURE!!
+"""
+
+from __future__ import annotations
+
+import typing as t
+
+
+class Mock:
+    """Base mock class"""
+
+    def __init__(self, *_: t.Any, **__: t.Any): ...
+    def __getattr__(self, _: str) -> Mock:
+        return Mock()
+
+
+# Remove with merge of #603
+# https://github.com/CrayLabs/SmartSim/pull/603
+class Job(Mock): ...
+
+
+# Remove with merge of #599
+# https://github.com/CrayLabs/SmartSim/pull/599
+class JobGroup(Mock): ...
+
+
+# Remove with merge of #587
+# https://github.com/CrayLabs/SmartSim/pull/587
+class LaunchSettings(Mock): ...
