+++ smartsim/entity/entity.py
@@ -24,52 +24,14 @@
 # OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-
 from __future__ import annotations
 
 import abc
-import copy
-import itertools
-import os
 import typing as t
-
-from smartsim.entity import strategies
-from smartsim.entity.files import EntityFiles
-from smartsim.entity.model import Application
 
 if t.TYPE_CHECKING:
     import smartsim.settings.base
-
-
-# >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>
-# TODO: Mocks to be removed later
-# ------------------------------------------------------------------------------
-# pylint: disable=multiple-statements
-
-
-class _Mock:
-    def __init__(self, *_: t.Any, **__: t.Any): ...
-    def __getattr__(self, _: str) -> "_Mock":
-        return _Mock()
-
-
-# Remove with merge of #603
-# https://github.com/CrayLabs/SmartSim/pull/603
-class Job(_Mock): ...
-
-
-# Remove with merge of #599
-# https://github.com/CrayLabs/SmartSim/pull/599
-class JobGroup(_Mock): ...
-
-
-# Remove with merge of #587
-# https://github.com/CrayLabs/SmartSim/pull/587
-class LaunchSettings(_Mock): ...
-
-
-# pylint: enable=multiple-statements
-# <<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<
+    from smartsim.entity import _mock
 
 
 class TelemetryConfiguration:
@@ -163,58 +125,6 @@ class SmartSimEntity:
 
 class CompoundEntity(abc.ABC):
     @abc.abstractmethod
-    def as_jobs(self, settings: LaunchSettings) -> t.Collection[Job]: ...
-    def as_job_group(self, settings: LaunchSettings) -> JobGroup:
-        return JobGroup(self.as_jobs(settings))
-
-
-# TODO: If we like this design, we need to:
-#         1) Move this to the `smartsim._core.entity.ensemble` module
-#         2) Decide what to do witht the original `Ensemble` impl
-class Ensemble(CompoundEntity):
-    def __init__(
-        self,
-        name: str,
-        exe: str | os.PathLike[str],
-        exe_args: t.Sequence[str] | None = None,
-        files: EntityFiles | None = None,
-        parameters: t.Mapping[str, t.Sequence[str]] | None = None,
-        permutation_strategy: str | strategies.TPermutationStrategy = "all_perm",
-        max_permutations: int = 0,
-        replicas: int = 1,
-    ) -> None:
-        self.name = name
-        self.exe = os.fspath(exe)
-        self.exe_args = list(exe_args) if exe_args else []
-        self.files = copy.deepcopy(files) if files else EntityFiles()
-        self.parameters = dict(parameters) if parameters else {}
-        self.permutation_strategy = permutation_strategy
-        self.max_permutations = max_permutations
-        self.replicas = replicas
-
-    def _create_applications(self) -> tuple[Application, ...]:
-        permutation_strategy = strategies.resolve(self.permutation_strategy)
-        permutations = permutation_strategy(self.parameters, self.max_permutations)
-        permutations = permutations if permutations else [{}]
-        permutations_ = itertools.chain.from_iterable(
-            itertools.repeat(permutation, self.replicas) for permutation in permutations
-        )
-        return tuple(
-            Application(
-                name=f"{self.name}-{i}",
-                exe=self.exe,
-                run_settings=_Mock(),  # type: ignore[arg-type]
-                # ^^^^^^^^^^^^^^^^^^^
-                # FIXME: remove this constructor arg! It should not exist!!
-                exe_args=self.exe_args,
-                files=self.files,
-                params=permutation,
-            )
-            for i, permutation in enumerate(permutations_)
-        )
-
-    def as_jobs(self, settings: LaunchSettings) -> tuple[Job, ...]:
-        apps = self._create_applications()
-        if not apps:
-            raise ValueError("There are no members as part of this ensemble")
-        return tuple(Job(app, settings) for app in apps)
+    def as_jobs(self, settings: _mock.LaunchSettings) -> t.Collection[_mock.Job]: ...
+    def as_job_group(self, settings: _mock.LaunchSettings) -> _mock.JobGroup:
+        return _mock.JobGroup(self.as_jobs(settings))
